# docs/swagger.yaml
basePath: /api/v1
definitions:
  dtos.ActivateWeekRequest:
    properties:
      week_number:
        type: integer
    required:
    - week_number
    type: object
...
  dtos.ContestWeekRequest:
    properties:
      end_date:
        type: string
      start_date:
        type: string
      week_number:
        minimum: 1
        type: integer
      winner_count:
        minimum: 1
        type: integer
    required:
    - end_date
    - start_date
    - week_number
    - winner_count
    type: object
  dtos.ContestWeekResponse:
    properties:
      created_on:
        type: string
      end_date:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      start_date:
        type: string
      week_number:
        type: integer
      winner_count:
        type: integer
    type: object
...
  dtos.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  dtos.ProfileResponseDTO:
    properties:
      user:
        $ref: '#/definitions/dtos.UserProfileDTO'
    type: object
...
  dtos.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dtos.UpdateProfileRequestDTO:
    properties:
      avatar_id:
        type: integer
      email:
        type: string
      name:
        type: string
      platform_user_name:
        type: string
      sharing_platform:
        type: string
    type: object
  dtos.UserProfileDTO:
    properties:
      avatar_image:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_verified:
        type: boolean
      name:
        type: string
      phone_number:
        type: string
      referral_code:
        type: string
      referred_by:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@thumsup.com
    name: API Support
  description: API documentation for Thums Up Backend Service including Contest Week
    Management
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Thums Up Backend API
  version: "1.0"
paths:
...
  /contest-weeks:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all contest weeks
      produces:
      - application/json
      responses:
        "200":
          description: Contest weeks retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtos.ContestWeekResponse'
                  type: array
              type: object
        "500":
          description: Failed to get contest weeks
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get all contest weeks
      tags:
      - Contest Weeks
    post:
      consumes:
      - application/json
      description: Create a new contest week with week number, start date, end date,
        and winner count. Requires authentication.
      parameters:
      - description: Contest week details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ContestWeekRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Contest week created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.ContestWeekResponse'
              type: object
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to create contest week
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new contest week
      tags:
      - Contest Weeks
  /contest-weeks/{weekNumber}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific contest week by its week number
      parameters:
      - description: Week number
        in: path
        name: weekNumber
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Contest week retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.ContestWeekResponse'
              type: object
        "400":
          description: Invalid week number
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Contest week not found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to get contest week
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get contest week by week number
      tags:
      - Contest Weeks
  /contest-weeks/activate:
    post:
      consumes:
      - application/json
      description: Activate a specific contest week by week number. Only one week
        can be active at a time. Requires authentication.
      parameters:
      - description: Week number to activate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ActivateWeekRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Contest week activated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.ContestWeekResponse'
              type: object
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Contest week not found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to activate contest week
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - Bearer: []
      summary: Activate a contest week
      tags:
      - Contest Weeks
  /contest-weeks/active:
    get:
      consumes:
      - application/json
      description: Retrieve the currently active contest week
      produces:
      - application/json
      responses:
        "200":
          description: Active week retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.ContestWeekResponse'
              type: object
        "404":
          description: No active contest week found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to get active week
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get active contest week
      tags:
      - Contest Weeks
  /profile:
    get:
      consumes:
      - application/json
      description: Retrieve the authenticated user's profile information. Requires
        authentication.
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved successfully
          schema:
            $ref: '#/definitions/dtos.ProfileResponseDTO'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch profile
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get user profile
      tags:
      - Profile
    patch:
      consumes:
      - application/json
      description: Update the authenticated user's profile information (name, email,
        avatar). Requires authentication.
      parameters:
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateProfileRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/dtos.UserProfileDTO'
        "400":
          description: Validation failed or email already in use
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update profile
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update user profile
      tags:
      - Profile
...
securityDefinitions:
  APIKey:
    description: Admin API key for protected operations
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"